Shading Start
Start to the shading aspect of project 3.

I moved to assimp for handling model loading in addition to the other nice features assimp offers.

Currently each face is being drawn properly. With proper vertex duplication in thanks to assimp.

As a test each vertex has a color from its normal. The abs of a normal value is used per vertex & if no normal values are found then random colors are created per vertex. This is temp just to visualize the models better.